requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
drf-yasg==1.17.1

